% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bolasso.R
\name{bootstrap_lasso}
\alias{bootstrap_lasso}
\title{bootstrap_lasso}
\usage{
bootstrap_lasso(x,y,bs=10,kfold=10,family,standardize)
}
\arguments{
\item{x}{Dataframe, containing the features of interest.}

\item{y}{Vector, containing status or the response.}

\item{bs}{Integer, number of boostrap.}

\item{kfold}{Integer, K-Folds cross-validator. Split dataset
into k consecutive folds.}

\item{family}{String, currently two options cox or binomial,
not all model available.}

\item{standardize}{Boolean, columns of the data matrix x are standardized,
i.e. each column of x has mean 0 and standard deviation 1.}
}
\description{
Generate a list of selected covariates for each bootstrap applied to a
logistic regression or cox lasso. (inspire from SparseLearner package)
}
