% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyMRDataSim_polynomial_function.R
\name{get_polynomial_function}
\alias{get_polynomial_function}
\title{Function factory to create polynomial functions}
\usage{
get_polynomial_function(polynomial_coefficients = c(0.1))
}
\arguments{
\item{polynomial_coefficients}{A vector of coefficients for the polynomial
terms, in ascending order of degree (default is c(0.1)). The length of the
vector determines the degree of the resulting polynomial.}
}
\value{
A polynomial function with the corresponding coefficients.
}
\description{
This function factory is intended for use with the PolyMRDataSim class. It
simplifies the creation of polynomial functions based on its coefficients.
}
\examples{
  cubic_polynomial <- get_polynomial_function(c(0.5, 0, 2))

  some_values <- rnorm(100)
  all(cubic_polynomial(some_values) == 0.5*some_values + 2*some_values^3)

}
