% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyMRDataSim.R
\name{new_PolyMRDataSim}
\alias{new_PolyMRDataSim}
\title{Simulate exposure, outcome, and genotype data}
\usage{
new_PolyMRDataSim(
  sample_size = 1e+05,
  n_exposure_snps = 100,
  exposure_heritability = 0.3,
  causal_function = get_polynomial_function(c(0.1, 0.05)),
  confounders_list = list(new_Confounder(sample_size)),
  finalize = TRUE,
  gws_thr = 5e-08
)
}
\arguments{
\item{sample_size}{Sample size (default is 10^5).}

\item{n_exposure_snps}{Number of SNPs explaining the
\code{exposure_heritability} (default is 100). Note that this is not equal
to the number of instruments as it includes SNPs which will be filtered out
upon finalizing the data (see \code{gws_thr}).}

\item{exposure_heritability}{Heritability of the exposure explained by the
\code{n_exposure_snps} (default is 0.3).}

\item{causal_function}{Function defining the true relationship between the
exposure and the outcome. It should accept a vector of exposure values and
return a vector of outcome values of the same length. This represents the
pure contribution of the exposure to the outcome and should not include
confounding or noise. Default is \code{function(x) 0.1*x + 0.05*x^2}. See
also [get_polynomial_function()]}

\item{confounders_list}{A list of objects of class \code{Confounder} (see
[new_Confounder()]) to be added to the data. Default is a single confounder
with linear contributions to both exposure and outcome with coefficients
0.2 and 0.5, respectively.}

\item{finalize}{Logical indicating whether the data set should be finalized,
i.e. errors added to contribute remaining variance and SNPs filtered based
on genome-wide significance (default is TRUE).}

\item{gws_thr}{P-value genome-wide significance threshold to filter SNPs for
instrumental variable selection (default is 5e-8).}
}
\value{
A list-like object of class \code{PolyMRDataSim}. It's main
  constituents are the \code{exposure} and \code{outcome} vectors, and the
  \code{genotypes} matrix. In addition, a number of parameters used in data
  generation are kept as named elements, including \code{n_exposure_snps},
  \code{exposure_heritability}, and \code{causal_function}. Some intermediate
  values are also included, namely the minor allele frequencies (\code{mafs})
  and the effects on the exposure (\code{exposure_coefficients}) of the SNPs
  remaining after filtering. These represent the ground-truth values used in
  the data generation.
}
\description{
Simulates exposure, outcome, and genotype data corresponding to the provided
causal function.
}
\examples{
  simulated_data <- new_PolyMRDataSim(
    sample_size = 50000,
    n_exposure_snps = 200,
    causal_function = function(x) 0.05*exp(x))

}
