% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polymr.R
\name{polymr}
\alias{polymr}
\title{Perform polynomial Mendelian}
\usage{
polymr(
  x,
  y,
  G,
  others = c("model", "means", "offset"),
  bins = 100,
  reverse_t = NULL,
  ...
)
}
\arguments{
\item{x}{A vector containing the individual exposure values}

\item{y}{A vector containing the individual outcome values}

\item{G}{The genetic matrix}

\item{others}{Character vector containing other values to return.
'model' returns the polynomial approximation of the observed association,
'offset' returns the scaling parameters of x and y,
'means' returns the mean value of y for each bin while 'medians' returns their median.}

\item{bins}{Number of bins for which to return median or mean values}

\item{reverse_t}{Threshold to use for reverse causality filtering (T statistic), NULL for no filtering}

\item{...}{Other parameters to pass to the non-linear methods}
}
\value{
List of results for PolyMR and the other selected values.
    'polymr' is a list containing 'pval0' (the p-value of the full model vs. no causal effect),
    'pval1' (the p-value for non-linearity),
    'r.squared' (the difference in explained variance full model vs. no causal effect),
    'coefficients' (the alpha parameter estimates and standard errors), and
    'vcov' (the variance-covariance matrix of the parameter estimates).
}
\description{
Perform polynomial Mendelian
}
