% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly.R
\name{polynomial_features}
\alias{polynomial_features}
\title{polynomial_features}
\usage{
polynomial_features(df,degree)
}
\arguments{
\item{df}{Dataframe, containing the features of interest.}

\item{degree}{Integer, the degree of the polynomial.}
}
\description{
Generate a new feature matrix consisting of all polynomial
combinations of the features with degree less than or equal to the specified
degree.
}
\examples{
val1 = rnorm(50)
val2 = rnorm(50)
df = as.data.frame(cbind(val1,val2))
df_poly = polynomial_features(df,2)
}
